test: add more tests
